# apps/rag-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LangConnect API",
    "description": "A REST API for a RAG system using FastAPI and LangChain",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Collections List",
        "description": "Lists all available vector collections (name and UUID).",
        "operationId": "collections_list_api_v1_collections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  },
                  "type": "array",
                  "title": "Response Collections List Api V1 Collections Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Collections Create",
        "description": "Creates a new vector collection by name with optional metadata.",
        "operationId": "collections_create_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreate"
              },
              "example": {
                "name": "Sample Name",
                "metadata": {
                  "key": "value",
                  "nested": {
                    "data": "example"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/collections/{collection_id}": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Collections Get",
        "description": "Retrieves details (name and UUID) of a specific vector collection.",
        "operationId": "collections_get_api_v1_collections__collection_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Collections Delete",
        "description": "Deletes a specific vector collection by name.",
        "operationId": "collections_delete_api_v1_collections__collection_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "collections"
        ],
        "summary": "Collections Update",
        "description": "Updates a specific vector collection's name and/or metadata.",
        "operationId": "collections_update_api_v1_collections__collection_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdate"
              },
              "example": {
                "name": "Sample Name",
                "metadata": {
                  "key": "value",
                  "nested": {
                    "data": "example"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/graph/build": {
      "post": {
        "tags": [
          "graph-rag"
        ],
        "summary": "Build Graph",
        "description": "Trigger knowledge graph construction from a vector collection.\n\nThis runs in the background and returns immediately.\nPoll /graph/build/{collection_id}/status to track progress.",
        "operationId": "build_graph_api_v1_graph_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphBuildRequest"
              },
              "example": {
                "collection_id": "abc123def456",
                "entity_types": "user",
                "relationship_types": "user"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/graph/build/{collection_id}/status": {
      "get": {
        "tags": [
          "graph-rag"
        ],
        "summary": "Get Build Status",
        "description": "Get the current build progress for a collection.\n\nReturns ``null`` when no build has ever been started for this\ncollection, so the frontend can distinguish \"never built\" from\n\"pending\".",
        "operationId": "get_build_status_api_v1_graph_build__collection_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BuildProgress"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Build Status Api V1 Graph Build  Collection Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/build/{collection_id}/pause": {
      "post": {
        "tags": [
          "graph-rag"
        ],
        "summary": "Pause Build",
        "description": "Pause a currently running graph build.",
        "operationId": "pause_build_api_v1_graph_build__collection_id__pause_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/build/{collection_id}/resume": {
      "post": {
        "tags": [
          "graph-rag"
        ],
        "summary": "Resume Build",
        "description": "Resume a paused graph build.",
        "operationId": "resume_build_api_v1_graph_build__collection_id__resume_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/build/{collection_id}/stop": {
      "post": {
        "tags": [
          "graph-rag"
        ],
        "summary": "Stop Build",
        "description": "Request cancellation for a currently running graph build.",
        "operationId": "stop_build_api_v1_graph_build__collection_id__stop_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphBuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/collections": {
      "get": {
        "tags": [
          "graph-rag"
        ],
        "summary": "List Graph Collections",
        "description": "Return the list of collection IDs that have a knowledge graph built.",
        "operationId": "list_graph_collections_api_v1_graph_collections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/graph/collections/{collection_id}/edges": {
      "get": {
        "tags": [
          "graph-rag"
        ],
        "summary": "List Edges",
        "description": "List edges (relationships) in the knowledge graph for a collection.",
        "operationId": "list_edges_api_v1_graph_collections__collection_id__edges_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The unique name of the collection.",
            "example": "Sample Name"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Optional metadata for the collection.",
            "example": {
              "key": "value",
              "nested": {
                "data": "example"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CollectionCreate",
        "description": "Schema for creating a new collection."
      },
      "CollectionResponse": {
        "properties": {
          "uuid": {
            "type": "string",
            "title": "Uuid",
            "description": "The unique identifier of the vector collection.",
            "example": "abc123def456"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the collection.",
            "example": "Sample Name"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Metadata associated with the collection.",
            "example": {
              "key": "value",
              "nested": {
                "data": "example"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "name"
        ],
        "title": "CollectionResponse",
        "description": "Schema for representing a vector collection."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}